requirements.txt
flask
flask-cors
flask-login
bcrypt
cffi
pymysql
sqlalchemy
# 其他依赖... 